backend/requirements.txt
Django>=3.2
djangorestframework
gunicorn
django-cors-headers
python-dotenv
cloudinary
django-cloudinary-storage